+++ debug_toolbar/panels/request.py
@@ -3,7 +3,11 @@ from django.urls import resolve
 from django.utils.translation import gettext_lazy as _
 
 from debug_toolbar.panels import Panel
-from debug_toolbar.utils import get_name_from_obj, get_sorted_request_variable
+from debug_toolbar.utils import (
+    get_name_from_obj,
+    get_sorted_request_variable,
+    sanitize_value,
+)
 
 
 class RequestPanel(Panel):
@@ -24,6 +28,8 @@ class RequestPanel(Panel):
         return view_func.rsplit(".", 1)[-1]
 
     def generate_stats(self, request, response):
+        from debug_toolbar import settings as dt_settings
+
         self.record_stats(
             {
                 "get": get_sorted_request_variable(request.GET),
@@ -58,14 +64,35 @@ class RequestPanel(Panel):
             pass
         self.record_stats(view_info)
 
+        # Handle session data with sanitization
         if hasattr(request, "session"):
+            sanitize_request_data = dt_settings.get_config().get(
+                "SANITIZE_REQUEST_DATA", True
+            )
+
             try:
-                session_list = [
-                    (k, request.session.get(k)) for k in sorted(request.session.keys())
-                ]
+                if sanitize_request_data:
+                    session_list = []
+                    for k in sorted(request.session.keys()):
+                        v = request.session.get(k)
+                        sanitized_v = sanitize_value(k, v)
+                        session_list.append((k, sanitized_v))
+                else:
+                    session_list = [
+                        (k, request.session.get(k))
+                        for k in sorted(request.session.keys())
+                    ]
             except TypeError:
-                session_list = [
-                    (k, request.session.get(k))
-                    for k in request.session.keys()  # (it's not a dict)
-                ]
+                # Handle non-dict session objects
+                if sanitize_request_data:
+                    session_list = []
+                    for k in request.session.keys():
+                        v = request.session.get(k)
+                        sanitized_v = sanitize_value(k, v)
+                        session_list.append((k, sanitized_v))
+                else:
+                    session_list = [
+                        (k, request.session.get(k)) for k in request.session.keys()
+                    ]
+
             self.record_stats({"session": {"list": session_list}})